Moved DefaultTags to SentryOptions.
