+++ src/Sentry.Extensions.Logging/SentryLoggingOptionsExtensions.cs
@@ -37,18 +37,6 @@ namespace Sentry.Extensions.Logging
             Func<string, LogLevel, EventId, Exception?, bool> filter)
             => options.AddLogEntryFilter(new DelegateLogEntryFilter(filter));
 
-        /// <summary>
-        /// Applies the default tags to an event without resetting existing tags.
-        /// </summary>
-        /// <param name="options">The options to read the default tags from.</param>
-        /// <param name="event">The event to apply the tags to.</param>
-        public static void ApplyDefaultTags(this SentryLoggingOptions options, SentryEvent @event)
-        {
-            foreach (var defaultTag in options.DefaultTags
-                .Where(t => !@event.Tags.TryGetValue(t.Key, out _)))
-            {
-                @event.SetTag(defaultTag.Key, defaultTag.Value);
-            }
-        }
+        
     }
 }
